EMG/hardware/requirements.txt
joblib==1.2.0
numpy==1.24.3
paho_mqtt==2.1.0
pyserial==3.5
